Create filter method
